+++ src/pages/PostsManagerPage.tsx
@@ -1,19 +1,21 @@
 import { useEffect } from "react"
 import { Plus } from "lucide-react"
-import { Button, Card, CardContent, CardHeader, CardTitle } from "../components"
-import { PostTable } from "../entities/post/ui/PostTable"
-import { SearchBar } from "../entities/post/ui/SearchBar"
-import { TagFilter } from "../entities/tag/ui/TagFilter"
-import { SortFilter } from "../features/filterPosts/ui/SortFilter"
-import { Pagination } from "../shared/ui/Pagination"
-import { AddPostDialog } from "../entities/post/ui/AddPostDialog"
-import { EditPostDialog } from "../entities/post/ui/EditPostDialog"
-import { PostDetailDialog } from "../entities/post/ui/PostDetailDialog"
-import { AddCommentDialog } from "../entities/comment/ui/AddCommentDialog"
-import { EditCommentDialog } from "../entities/comment/ui/EditCommentDialog"
-import { UserModal } from "../entities/user/ui/UserModal"
-import { usePostsManager } from "../features/postManagement/model/usePostsManager"
+
 import { usePostStore } from "../entities/post/model/store"
+import { Card, CardContent, CardHeader, CardTitle } from "../shared/ui/Card"
+import { Button } from "../shared/ui/Button"
+import { SortFilter } from "../features.tsx/filterPosts/ui/SortFilter"
+import Pagination from "../shared/ui/Pagenation"
+import { usePostsManager } from "../features.tsx/postManagement/model/usePostsManager"
+import AddPostDialog from "../entities/post/ui/AddPostDialog"
+import EditPostDialog from "../entities/post/ui/EditPostDialog"
+import PostDetailDialog from "../entities/post/ui/PostDetailDialog"
+import AddCommentDialog from "../entities/comment/ui/AddCommentDialog"
+import EditCommentDialog from "../entities/comment/ui/EditCommentDialog"
+import SearchBar from "../entities/post/ui/SearchBar"
+import TagFilter from "../entities/tag/ui/TagFilter"
+import UserModal from "../entities/user/ui/UserModal"
+import PostTable from "../entities/post/ui/PostTable"
 
 const PostsManagerPage = () => {
   const { loading } = usePostStore()
